+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 - Tumble Behind
+- Mistform Elixir
+- Elixir of Life
 
 ### Changed